requirements.txt
python-dotenv
requests
pandas
gspread
oauth2client
psycopg2-binary
sqlalchemy
pytest